Avoid using Facter values that may not be available on first run
This module defines facts that may change as a result of applying resources that
it also defines. Changing behaviour based on these facts is not correct, because
Puppet gathers all facts before compiling the resource catalog, so the behaviour
becomes non-deterministic. This is an haphazard patch that crudely avoids this
pitfall, by hard-coding the correct values for its author’s specific environment.